Arrays/List
Created a one-dimensional Array of strings and then one of integers. Also created a list of strings. They all ask the user to select an index of the Array and then display the integer/string.
There is also a message that displays when the user selects an index that doesn’t exist.