Parsing Enums
Created an enum for the days of the week.

Prompted the user to enter the current day of the week.

Assigned the value to a variable of that enum data type and parsed through the user input

Wraped in a try/catch block and it prints "Please enter an actual day of the week.” to the console if an error occurs.